# Speed up BenchmarkPostings_Stats